+++ include/Zeni/Vertex3f.hxx
@@ -45,8 +45,8 @@ namespace Zeni {
     return const_cast<Point3f *>(&m_position);
   }
 
-  unsigned long Vertex3f::get_offset() const {
-    return reinterpret_cast<unsigned long>(&m_position) - reinterpret_cast<unsigned long>(this);
+  unsigned long long Vertex3f::get_offset() const {
+    return reinterpret_cast<unsigned long long>(&m_position) - reinterpret_cast<unsigned long long>(this);
   }
 
   const Uint32 & Vertex3f_Color::get_color() const {
